+++ src/V1/ServiceBricks.Cache.Model/Model/CacheModelConstants.cs
@@ -8,6 +8,11 @@
         /// <summary>
         /// Application settings keys for client API configuration.
         /// </summary>
-        public const string APPSETTING_CLIENT_APICONFIG = @"ServiceBricks:Cache:Client:Api";
+        public const string APPSETTING_CLIENT_APICONFIG = @"ServiceBricks:Cache:Client:Api";        
+
+        /// <summary>
+        /// Application settings keys for semaphore options.
+        /// </summary>
+        public const string APPSETTING_SEMAPHORE_OPTIONS = "ServiceBricks:Cache:Semaphore";
     }
 }